refactor: slice src/index.ts (2223 lines) into modular structure
- Extract templates: html.ts, sidebar.ts, navbar.ts, footer.ts, common-scripts.ts
- Extract page handlers: auth.ts, dashboard.ts, pos.ts, admin.ts, menu.ts, tables.ts, orders.ts, products.ts
- Reduce src/index.ts from 2223 to 39 lines (entry point only)
- Delete unused src/components/ directory (Layout, Navbar, Sidebar, UI components)
