--- src/components/layout/Sidebar.tsx
@@ -1,48 +1,0 @@
-export function Sidebar({ currentPath, user }) {
-  const menuItems = [
-    { path: '/', label: 'Dashboard', icon: 'grid' },
-    { path: '/pos', label: 'POS', icon: 'box' },
-    { path: '/menu', label: 'Menu', icon: 'file-text' },
-    { path: '/tables', label: 'Meja', icon: 'calendar' },
-    { path: '/orders', label: 'Pesanan', icon: 'shopping-cart' },
-  ];
-
-  const icons = {
-    grid: '<svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2"><rect x="3" y="3" width="7" height="7"></rect><rect x="14" y="3" width="7" height="7"></rect><rect x="14" y="14" width="7" height="7"></rect><rect x="3" y="14" width="7" height="7"></rect></svg>',
-    box: '<svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2"><path d="M12 2L2 7l10 5 10-5-10-5z"></path><path d="M2 17l10 5 10-5"></path><path d="M2 12l10 5 10-5"></path></svg>',
-    'file-text': '<svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2"><path d="M14 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V8z"></path><polyline points="14 2 14 8 20 8"></polyline><line x1="16" y1="13" x2="8" y2="13"></line><line x1="16" y1="17" x2="8" y2="17"></line></svg>',
-    calendar: '<svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2"><rect x="3" y="4" width="18" height="18" rx="2" ry="2"></rect><line x1="3" y1="10" x2="21" y2="10"></line></svg>',
-    'shopping-cart': '<svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2"><path d="M6 2L3 6v14a2 2 0 0 0 2 2h14a2 2 0 0 0 2-2V6l-3-4z"></path><line x1="3" y1="6" x2="21" y2="6"></line><path d="M16 10a4 4 0 0 1-8 0"></path></svg>',
-  };
-
-  return `
-    <aside class="sidebar">
-      <div class="sidebar-header">
-        <div class="sidebar-logo">
-          <svg width="24" height="24" viewBox="0 0 24 24" fill="currentColor">
-            <path d="M19 3H5c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h14c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2zm-7 14l-5-5 1.41-1.41L12 14.17l4.59-4.58L18 11l-6 6z"/>
-          </svg>
-          POS App
-        </div>
-      </div>
-      <nav class="sidebar-nav">
-        <ul class="sidebar-menu">
-          ${menuItems.map(item => `
-            <li class="sidebar-menu-item">
-              <a href="${item.path}" class="sidebar-menu-link ${currentPath === item.path ? 'active' : ''}">
-                ${icons[item.icon]}
-                ${item.label}
-              </a>
-            </li>
-          `).join('')}
-        </ul>
-      </nav>
-      <div class="sidebar-footer">
-        <div class="navbar-user" onclick="logout()" style="cursor: pointer;">
-          <svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2"><path d="M9 21H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h4"></path><polyline points="16 17 21 12 16 7"></polyline><line x1="21" y1="12" x2="9" y2="12"></line></svg>
-          <span>Logout${user ? ` (${user.name})` : ''}</span>
-        </div>
-      </div>
-    </aside>
-  `;
-}
