--- src/components/ui/Button.tsx
@@ -1,3 +1,0 @@
-export function Button({ children, variant = 'primary', onClick, type = 'button' }) {
-  return `<button class="btn btn-${variant}" type="${type}" onclick="${onClick || ''}">${children}</button>`;
-}
