--- src/components/ui/Table.tsx
@@ -1,20 +1,0 @@
-export function Table({ headers, rows }) {
-  return `
-    <div class="table-container">
-      <table class="table">
-        <thead>
-          <tr>
-            ${headers.map(h => `<th>${h}</th>`).join('')}
-          </tr>
-        </thead>
-        <tbody>
-          ${rows.map(row => `
-            <tr>
-              ${row.map(cell => `<td>${cell}</td>`).join('')}
-            </tr>
-          `).join('')}
-        </tbody>
-      </table>
-    </div>
-  `;
-}
